split up state into multiple files
